refactor: 불필요한 toInt 제거, 사영되지 않는 라이브데이터 제거, val 키워드 제거

+++ android/app/src/main/java/com/daedan/festabook/presentation/placeList/behavior/PlaceListBottomSheetBehavior.kt
@@ -50,7 +50,7 @@ class PlaceListBottomSheetBehavior<V : View>(
         recyclerView = child.findViewById<RecyclerView>(R.id.rv_places)
         expandedOffset = parent.findViewById<ChipGroup>(R.id.cg_categories).height
         recyclerView.getChildAt(HEADER_POSITION)?.let {
-            headerRange = expandedOffset.toInt()..(expandedOffset.toInt() + it.height)
+            headerRange = expandedOffset..(expandedOffset + it.height)
         }
         return super.onLayoutChild(parent, child, layoutDirection)
     }
